requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
jinja2
python-multipart
markdown
python-dotenv
loguru
passlib[bcrypt]
fastapi-csrf-protect
email-validator
argon2-cffi
fastapi-limiter
slowapi
